调整一下注册模块监听方法的调用位置，增强代码的安全性。
